requirements/dev.txt
-r common.txt
-r test.txt
django-debug-toolbar
django-debug-toolbar-request-history
django-debugtools
django-sslserver
django-developer-panel
